+++ project/src/main/java/org/example/fourtreesproject/emailVerify/model/entity/EmailVerify.java
@@ -1,9 +1,6 @@
 package org.example.fourtreesproject.emailVerify.model.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,6 +8,7 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor
+@Table(name="email_verify")
 @AllArgsConstructor
 @Getter
 @Builder
