+++ project/src/main/java/org/example/fourtreesproject/jwt/model/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package org.example.fourtreesproject.jwt;
+package org.example.fourtreesproject.jwt.model.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Builder
+@Table(name="refresh_token")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
